fix: use ptrdiff_t to fix MSVC build error

+++ pandas/_libs/src/parser/tokenizer.c
@@ -24,6 +24,7 @@ GitHub. See Python Software Foundation License and BSD licenses for these.
 #include <limits.h>
 #include <math.h>
 #include <stdbool.h>
+#include <stddef.h>
 #include <stdlib.h>
 
 #include "pandas/portable.h"
@@ -1950,8 +1951,8 @@ int64_t str_to_int64(const char *p_item, int64_t int_min, int64_t int_max,
 
     char *new_end = NULL;
     int64_t next_part = strtoll(endptr, &new_end, 10);
-    int digits = new_end - endptr;
-    int64_t mul_result = power_int(10, digits);
+    ptrdiff_t digits = new_end - endptr;
+    int64_t mul_result = power_int(10, (int)digits);
     result = add_int_check_overflow(result, next_part, mul_result);
     endptr = new_end;
   }
@@ -2008,8 +2009,8 @@ uint64_t str_to_uint64(uint_state *state, const char *p_item, int64_t int_max,
 
     char *new_end = NULL;
     uint64_t next_part = strtoull(endptr, &new_end, 10);
-    int digits = new_end - endptr;
-    uint64_t mul_result = power_int(10, digits);
+    ptrdiff_t digits = new_end - endptr;
+    uint64_t mul_result = power_int(10, (int)digits);
     result = add_uint_check_overflow(result, next_part, mul_result);
     endptr = new_end;
   }
